+++ Apptentive/Apptentive/Engagement/Interactions/Rating Flow/ApptentiveInteractionAppStoreController.m
@@ -11,6 +11,8 @@
 #import "ApptentiveUtilities.h"
 #import "ApptentiveInteraction.h"
 #import "ApptentiveBackend+Engagement.h"
+#import "UIAlertController+Apptentive.h"
+#import "ApptentiveStoreProductViewController.h"
 
 NSString *const ATInteractionAppStoreRatingEventLabelLaunch = @"launch";
 NSString *const ATInteractionAppStoreRatingEventLabelOpenAppStoreURL = @"open_app_store_url";
@@ -85,7 +87,11 @@ NSString *const ATInteractionAppStoreRatingEventLabelUnableToRate = @"unable_to_
 	UIAlertController *alertController = [UIAlertController alertControllerWithTitle:title message:message preferredStyle:UIAlertControllerStyleAlert];
 	[alertController addAction:[UIAlertAction actionWithTitle:cancelButtonTitle style:UIAlertActionStyleCancel handler:nil]];
 
-	[self.presentingViewController presentViewController:alertController animated:YES completion:nil];
+	if (self.presentingViewController != nil) {
+		[self.presentingViewController presentViewController:alertController animated:YES completion:nil];
+	} else {
+		[alertController apptentive_presentAnimated:YES completion:nil];
+	}
 }
 
 - (BOOL)shouldOpenAppStoreViaStoreKit {
@@ -139,7 +145,7 @@ NSString *const ATInteractionAppStoreRatingEventLabelUnableToRate = @"unable_to_
 
 - (void)openAppStoreViaStoreKit {
 	if ([SKStoreProductViewController class] != NULL && [self appID]) {
-		SKStoreProductViewController *vc = [[SKStoreProductViewController alloc] init];
+		ApptentiveStoreProductViewController *vc = [[ApptentiveStoreProductViewController alloc] init];
 		vc.delegate = self;
 		[vc loadProductWithParameters:@{ SKStoreProductParameterITunesItemIdentifier: self.appID } completionBlock:^(BOOL result, NSError *error) {
 			if (error) {
@@ -150,10 +156,10 @@ NSString *const ATInteractionAppStoreRatingEventLabelUnableToRate = @"unable_to_
 				
 				UIViewController *presentingVC = self.presentingViewController;
 
-				if (!presentingVC) {
-					ApptentiveLogError(@"Attempting to open the App Store via StoreKit from a nil View Controller!");
+				if (presentingVC != nil) {
+					[presentingVC presentViewController:vc animated:YES completion:nil];
 				} else {
-					[presentingVC presentViewController:vc animated:YES completion:^{}];
+					[vc presentAnimated:YES completion:nil];
 				}
 			}
 		}];